* CMYK color adjustments don't work. Noted this in TODO * Added todo about issues with adjustment layers. * Added cheap and cheerful thread pool implementation * Made filters listview use threadpool, leading to added speed in popping up the dialog * Removed lots and lots of warnings * Moved the thumbnail method that produces paint devices to KisPaintDevice, preparatory   to using it in the preview widget proper.
svn path=/trunk/koffice/; revision=509916

+++ krita/ui/kis_cmb_idlist.cc
@@ -77,7 +77,7 @@ void KisCmbIDList::setCurrentText(const QString & s)
 
 void KisCmbIDList::slotIDActivated(int i)
 {
-    if (i > m_list.count()) return;
+    if ((uint)i > m_list.count()) return;
 
     emit activated(m_list[i]);
 
@@ -85,7 +85,7 @@ void KisCmbIDList::slotIDActivated(int i)
 
 void KisCmbIDList::slotIDHighlighted(int i)
 {
-    if (i > m_list.count()) return;
+    if ((uint)i > m_list.count()) return;
 
     emit highlighted(m_list[i]);
